requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiogram3-calendar==0.1.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
config==0.5.1
dnspython==2.6.1
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
magic-filter==1.0.12
motor==3.3.2
multidict==6.0.5
packaging==23.2
pluggy==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
pytest==8.0.2
pytest-asyncio==0.23.5
python-dotenv==1.0.1
typing_extensions==4.9.0
yarl==1.9.4
